# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Thundergate API
  description: |
    Thundergate — a real-time execution firewall between autonomous AI agents
    and external APIs. Provides rule-based evaluation, human-in-the-loop
    review, and tamper-evident audit logging.
  version: 0.1.0
  contact:
    name: Thundergate Team
servers:
  - url: http://localhost:3001
    description: Proxy (Agent-facing API)
  - url: http://localhost:3000
    description: Dashboard (Operator-facing API)
...
paths:
...
  /queue/{queueId}/decide:
    post:
      tags: [queue]
      summary: Submit operator decision on a queue item
      parameters:
        - name: queueId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision:
                  type: string
                  enum: [APPROVED, MODIFIED, REJECTED]
                notes:
                  type: string
                modifiedPayload:
                  type: object
      responses:
        '200':
          description: Decision recorded

  # ─── Dashboard API — Queue (port 3000) ──────────────────
  /api/queue:
    get:
      tags: [queue]
      summary: List HITL queue items
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, CLAIMED, APPROVED, MODIFIED, REJECTED, ESCALATED, EXPIRED]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Paginated queue items
...
  /api/queue/{id}:
    get:
      tags: [queue]
      summary: Get queue item detail with audit log context
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Queue item with audit log
        '404':
          description: Queue item not found
  /api/queue/{id}/claim:
    post:
      tags: [queue]
      summary: Claim a queue item for review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Item claimed successfully
        '409':
          description: Item already claimed
  /api/queue/{id}/decide:
    post:
      tags: [queue]
      summary: Submit decision on a claimed queue item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision:
                  type: string
                  enum: [APPROVED, MODIFIED, REJECTED]
                notes:
                  type: string
                modifiedPayload:
                  type: object
      responses:
        '200':
          description: Decision recorded
...
  /api/rules:
    get:
      tags: [rules]
      summary: List all firewall rules
      parameters:
        - name: action
          in: query
          schema:
            type: string
            enum: [ALLOW, BLOCK, FLAG_FOR_REVIEW, MODIFY]
        - name: severity
          in: query
          schema:
            type: string
            enum: [LOW, MEDIUM, HIGH, CRITICAL]
        - name: enabled
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Array of rules
    post:
      tags: [rules]
      summary: Create a new rule
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleInput'
      responses:
        '201':
          description: Rule created
  /api/rules/{id}:
    get:
      tags: [rules]
      summary: Get rule detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Rule detail
    put:
      tags: [rules]
      summary: Update a rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleInput'
      responses:
        '200':
          description: Rule updated
  /api/rules/{id}/toggle:
    post:
      tags: [rules]
      summary: Enable or disable a rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                enabled:
                  type: boolean
      responses:
        '200':
          description: Rule status updated
...
  /api/targets:
    get:
      tags: [targets]
      summary: List API targets
      responses:
        '200':
          description: Array of targets
    post:
      tags: [targets]
      summary: Create a new API target
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TargetInput'
      responses:
        '201':
          description: Target created
  /api/targets/{id}:
    get:
      tags: [targets]
      summary: Get target detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Target detail
    put:
      tags: [targets]
      summary: Update a target
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Target updated
...
components:
  schemas:
...
    RuleInput:
      type: object
      required: [name, priority, conditions, action, severity]
      properties:
        name:
          type: string
        description:
          type: string
        priority:
          type: integer
          minimum: 1
          maximum: 1000
        conditions:
          type: object
          properties:
            url_pattern:
              type: string
            methods:
              type: array
              items:
                type: string
                enum: [GET, POST, PUT, PATCH, DELETE, HEAD, OPTIONS]
            payload_patterns:
              type: array
              items:
                type: string
            header_patterns:
              type: object
            agent_ids:
              type: array
              items:
                type: string
                format: uuid
            risk_score_threshold:
              type: integer
              minimum: 0
              maximum: 100
        action:
          type: string
          enum: [ALLOW, BLOCK, FLAG_FOR_REVIEW, MODIFY]
        severity:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        enabled:
          type: boolean
          default: true
...
    TargetInput:
      type: object
      required: [name, baseUrl]
      properties:
        name:
          type: string
        baseUrl:
          type: string
          format: uri
        riskTier:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
          default: MEDIUM
        headers:
          type: object
...
  securitySchemes:
    AgentApiKey:
      type: apiKey
      in: header
      name: X-Agent-Key
      description: Agent API key for proxy authentication
    SessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: Dashboard session cookie (NextAuth)